requirements-all.txt
# dev
pyinstaller==6.11.0
vermin==1.6.0
pillow==11.0.0

# app
pyzmq==26.1.1
Pygments==2.12.0
pygments-ansi-color==0.3.0
